remove automatic ordering of data by ID/stratum
We can assume that users have ordered the data correctly.
This is to avoid future bugs or unpredictable behaviour
where the package silently reorders the data without the
user's knowledge. In the future, we could possibly issue
an error message if observations for each track don't
appear to be contiguous.